requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
aiosqlite
gspread
aiofiles